+++ src/main/java/com/github/cbl/algorithm_analyzer/graphs/deepsearch/EdgeType.java
